# src/swaggerOutput.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "v1.0.0",
    "title": "Lazer Escort",
    "description": "Lazer Escort REST API Documentation"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": ""
    }
  ],
  "paths": {
    "/escort/login": {
      "post": {
        "tags": [
          "Escort Sign Up and Login"
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EscortLoginResponse"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/EscortLoginResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserNotFound"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/UserNotFound"
                }
              }
            }
          },
          "406": {
            "description": "Not Acceptable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidCredentials"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidCredentials"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        }
      }
    },
    "/escort/signup": {
      "post": {
        "tags": [
          "Escort Sign Up and Login"
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EscortLoginResponse"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/EscortLoginResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserExists"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/UserExists"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EscortSignupRequest"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/EscortSignupRequest"
              }
            }
          }
        }
      }
    },
...
    "/admin/packages": {
      "get": {
        "tags": [
          "Admin All Packages"
        ],
        "description": "",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PackagesResponse"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/PackagesResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidSession"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidSession"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserNotExists"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/UserNotExists"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PackagesRequest"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/PackagesRequest"
              }
            }
          }
        }
      }
    },
    "/admin/package": {
      "post": {
        "tags": [
          "Admin All Packages"
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Response"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidSession"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidSession"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserNotExists"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/UserNotExists"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PackageRequest"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/PackageRequest"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "LoginRequest": {
        "type": "object",
        "properties": {
          "usernameEmail": {
            "type": "string",
            "example": "string"
          },
          "password": {
            "type": "string",
            "example": "string"
          }
        },
        "required": [
          "usernameEmail",
          "password"
        ],
        "xml": {
          "name": "LoginRequest"
        }
      },
...
      "EscortSignupRequest": {
        "type": "object",
        "properties": {
          "workingName": {
            "type": "string",
            "example": "string"
          },
          "email": {
            "type": "string",
            "example": "string"
          },
          "password": {
            "type": "string",
            "example": "string"
          }
        },
        "required": [
          "workingName",
          "email",
          "password"
        ],
        "xml": {
          "name": "EscortSignupRequest"
        }
      },
...
      "PackageRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "string"
          },
          "expressLimit": {
            "type": "string",
            "example": "any"
          },
          "showLimit": {
            "type": "string",
            "example": "any"
          },
          "uploadLimit": {
            "type": "string",
            "example": "any"
          },
          "validityPeriod": {
            "type": "string",
            "example": "any"
          },
          "price": {
            "type": "string",
            "example": "number"
          }
        },
        "required": [
          "name",
          "expressLimit",
          "showLimit",
          "uploadLimit",
          "validityPeriod",
          "price"
        ],
        "xml": {
          "name": "PackageRequest"
        }
      },
...
      "PackagesRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "string"
          }
        },
        "xml": {
          "name": "PackagesRequest"
        }
      },
...
      "Response": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "string"
          }
        },
        "xml": {
          "name": "Response"
        }
      },
...
      "EscortLoginResponse": {
        "type": "object",
        "properties": {
          "workingName": {
            "type": "string",
            "example": "string"
          },
          "email": {
            "type": "string",
            "example": "string"
          },
          "token": {
            "type": "string",
            "example": "string"
          },
          "message": {
            "type": "string",
            "example": "string"
          },
          "profile": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "example": "string"
              },
              "workingName": {
                "type": "string",
                "example": "string"
              },
              "email": {
                "type": "string",
                "example": "string"
              },
              "verified": {
                "type": "string",
                "example": "boolean"
              },
              "password": {
                "type": "string",
                "example": "string"
              },
              "createdAt": {
                "type": "string",
                "example": "number"
              },
              "personalDetails": {
                "type": "object",
                "properties": {
                  "gender": {
                    "type": "string",
                    "example": "string"
                  },
                  "sexuality": {
                    "type": "string",
                    "example": "string"
                  },
                  "age": {
                    "type": "string",
                    "example": "number"
                  },
                  "nationality": {
                    "type": "string",
                    "example": "string"
                  }
                }
              },
              "physicalDetails": {
                "type": "object",
                "properties": {
                  "chest": {
                    "type": "string",
                    "example": "string"
                  },
                  "waist": {
                    "type": "string",
                    "example": "string"
                  },
                  "hips": {
                    "type": "string",
                    "example": "string"
                  },
                  "ethnicity": {
                    "type": "string",
                    "example": "string"
                  },
                  "hairColour": {
                    "type": "string",
                    "example": "string"
                  },
                  "height": {
                    "type": "string",
                    "example": "number"
                  },
                  "weight": {
                    "type": "string",
                    "example": "number"
                  },
                  "eyeColour": {
                    "type": "string",
                    "example": "string"
                  },
                  "genetalia": {
                    "type": "string",
                    "example": "string"
                  },
                  "cupSize": {
                    "type": "string",
                    "example": "string"
                  },
                  "breastImplant": {
                    "type": "string",
                    "example": "string"
                  },
                  "bodyType": {
                    "type": "string",
                    "example": "string"
                  },
                  "bodyArt": {
                    "type": "string",
                    "example": "string"
                  }
                }
              },
              "languages": {
                "type": "array",
                "example": [
                  "string"
                ],
                "items": {
                  "type": "string"
                }
              },
              "bookingNotes": {
                "type": "array",
                "example": [
                  "string"
                ],
                "items": {
                  "type": "string"
                }
              },
              "location": {
                "type": "object",
                "properties": {
                  "incall": {
                    "type": "string",
                    "example": "string"
                  },
                  "outcall": {
                    "type": "object",
                    "properties": {
                      "location": {
                        "type": "string",
                        "example": "string"
                      },
                      "iTravelTo": {
                        "type": "string",
                        "example": "string"
                      }
                    }
                  }
                }
              },
              "price": {
                "type": "object",
                "properties": {
                  "incall": {
                    "type": "object",
                    "properties": {
                      "hour1": {
                        "type": "string",
                        "example": "number"
                      },
                      "hour2": {
                        "type": "string",
                        "example": "number"
                      },
                      "hour3": {
                        "type": "string",
                        "example": "number"
                      }
                    }
                  },
                  "outcall": {
                    "type": "object",
                    "properties": {
                      "hour1": {
                        "type": "string",
                        "example": "number"
                      },
                      "hour2": {
                        "type": "string",
                        "example": "number"
                      },
                      "hour3": {
                        "type": "string",
                        "example": "number"
                      }
                    }
                  }
                }
              },
              "availability": {
                "type": "object",
                "properties": {
                  "monday": {
                    "type": "string",
                    "example": "boolean"
                  },
                  "tueday": {
                    "type": "string",
                    "example": "boolean"
                  },
                  "wednesday": {
                    "type": "string",
                    "example": "boolean"
                  },
                  "thurday": {
                    "type": "string",
                    "example": "boolean"
                  },
                  "friday": {
                    "type": "string",
                    "example": "boolean"
                  },
                  "saturday": {
                    "type": "string",
                    "example": "boolean"
                  },
                  "dunurday": {
                    "type": "string",
                    "example": "boolean"
                  }
                }
              },
              "services": {
                "type": "array",
                "example": [
                  "69",
                  "blowjob",
                  "couples services",
                  "dildo play",
                  "erotic massage",
                  "golden shower",
                  "hardcore domina",
                  "striptease",
                  "threesome"
                ],
                "items": {
                  "type": "string"
                }
              },
              "images": {
                "type": "array",
                "example": [
                  "string"
                ],
                "items": {
                  "type": "string"
                }
              },
              "videos": {
                "type": "array",
                "example": [
                  "string"
                ],
                "items": {
                  "type": "string"
                }
              }
            }
          }
        },
        "xml": {
          "name": "EscortLoginResponse"
        }
      },
...
      "PackagesResponse": {
        "type": "object",
        "properties": {
          "packages": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "string"
                },
                "expressLimit": {
                  "type": "string",
                  "example": "string"
                },
                "showLimit": {
                  "type": "string",
                  "example": "string"
                },
                "uploadLimit": {
                  "type": "string",
                  "example": "string"
                },
                "validityPeriod": {
                  "type": "string",
                  "example": "string"
                },
                "price": {
                  "type": "string",
                  "example": "number"
                }
              }
            }
          }
        },
        "xml": {
          "name": "PackagesResponse"
        }
      },
...
      "InvalidSession": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Invalid Session : Login again"
          }
        },
        "xml": {
          "name": "InvalidSession"
        }
      },
...
      "InvalidCredentials": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Invalid credentials"
          }
        },
        "xml": {
          "name": "InvalidCredentials"
        }
      },
      "UserNotFound": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "User not found"
          }
        },
        "xml": {
          "name": "UserNotFound"
        }
      },
      "UserNotExists": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "User does not exists"
          }
        },
        "xml": {
          "name": "UserNotExists"
        }
      },
...
      "UserExists": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "User already exists"
          }
        },
        "xml": {
          "name": "UserExists"
        }
      }
    }
  },
...
}